#FDF-128: Performance issues in lean memory changes TESTS_RAN: Unit test, dev test
git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2500 01a69087-22d5-4a29-8152-8a1d5e10e5e9

+++ ssd/fifo/mcd_rec.h
@@ -542,7 +542,7 @@ typedef struct mcd_rec_log {
 
     fthLock_t               slablock;         // access deferred slab info
     uint                    nslab;            // # deferred slab deallocs
-    uint32_t                slabtab[10000000]; // deferred-dealloc slabs
+    uint64_t                slabtab[10000000]; // deferred-dealloc slabs
     mlog_t                  mlog;
 
 	// lean memory extension
